--- a/.gitignore
+++ b/.gitignore
@@ -9,3 +9,4 @@
 /test/tmp
 package-lock.json
 /dist
+**/.DS_Store