--- a/.gitignore
+++ b/.gitignore
@@ -4,6 +4,7 @@
 /coverage
 /coverage.lcov
 /lib
+/dist
 /node_modules
 /tmp
 /test/tmp