--- a/.gitignore
+++ b/.gitignore
@@ -9,5 +9,4 @@
 /tmp
 /test/tmp
 package-lock.json
-/dist
 **/.DS_Store