--- a/src/log.ts
+++ b/src/log.ts
@@ -1,11 +1,11 @@
 import {CliUx} from '@oclif/core'
-import * as qq from 'qqjs'
 import * as util from 'util'
+import {prettifyPaths} from './util'
 
 export const debug = require('debug')('oclif')
 debug.new = (name: string) => require('debug')(`oclif:${name}`)
 
 export function log(format: string, ...args: any[]): void {
-  args = args.map((arg: any) => qq.prettifyPaths(arg))
+  args = args.map((arg: any) => prettifyPaths(arg))
   debug.enabled ? debug(format, ...args) : CliUx.ux.log(`oclif: ${util.format(format, ...args)}`)
 }