… .DS_Store
… .github/workflows/test.yml
… .vscode/launch.json
… package.json
… src/commands/upload/macos.ts
… src/commands/upload/tarballs.ts
… src/commands/upload/win.ts
… src/tarballs/bin.ts
… src/tarballs/build.ts
… src/tarballs/config.ts
… src/tarballs/node.ts
… src/upload-util.ts
… src/util.ts
… test/helpers/helper.ts
… test/integration/deb.test.ts
… test/integration/macos.test.ts
… test/integration/publish.test.ts
… test/integration/win.test.ts
… test/release.key
… tsconfig.json
… yarn.lock