we had to drop the --parallel flag on builds because of a bug.

remove `qqjs` and its `qq.cd` that's causing the problem with parallelization.

make sure each build process has its own directory isolation

parallelize all the things

@W-12107459@

### side effects
fix https://github.com/oclif/oclif/issues/347
test: restores integration testing for Debian builds (comment said they weren't runnable on CircleCI. Works on gha)
test: runs integration tests in parallel on different OSes to test both mac and deb packing


---
perf metrics [mac, 2.4 GHz 8-Core Intel Core i9]
1. clear all the cache files `rm -rf dist test/tmp tmp `
2. yarn test:integration

| test        | this PR (ms)           | main (ms) |
| ------------- |:-------------:| :-----:|
| macos | 86,942 | 109,495 |
| tarballs | 281,566  |  696,853  |
| win | 97,470 |  136,077   |

on gha builds (2 core machine).  [mac tests can't run because they only run on macs]

| test        | [this PR ](https://github.com/oclif/oclif/actions/runs/3527800462/jobs/5917250522)(ms)           | [main](https://github.com/oclif/oclif/actions/runs/3506189602/jobs/5873058319#step:9:1) (ms) |
| ------------- |:-------------:| :-----:|
| tarballs | 357,739  |  594,614  |
| win | 67,872 |  101,885   |