@mshanemc 2022-11-30
> Did you do any testing on the promote command? I know integration tests don't exist for it currently but might be nice to add some with all the changes going in.

just manually.   I checked it again with the new changes via.
```sh
../oclif-repos/oclif-oclif/bin/dev promote --channel qqqa --version 7.179.0 --sha 4383be1
```

I'm looking into the integration test options